+++ osu.Game/Screens/Multi/RealtimeMultiplayer/RealtimeMultiplayer.cs
@@ -5,6 +5,7 @@ using osu.Framework.Allocation;
 using osu.Framework.Logging;
 using osu.Framework.Screens;
 using osu.Game.Extensions;
+using osu.Game.Graphics.UserInterface;
 using osu.Game.Online.Multiplayer;
 using osu.Game.Online.RealtimeMultiplayer;
 using osu.Game.Screens.Multi.Components;
@@ -64,5 +65,7 @@ namespace osu.Game.Screens.Multi.RealtimeMultiplayer
         protected override RoomManager CreateRoomManager() => new RealtimeRoomManager();
 
         protected override LoungeSubScreen CreateLounge() => new RealtimeLoungeSubScreen();
+
+        protected override OsuButton CreateNewMultiplayerGameButton() => new CreateRealtimeMatchButton();
     }
 }
